pyproject.toml
[project]
name = "bq-streamlit"
version = "0.1.0"
description = "A test of streamlit to get data from BigQuery"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "black>=25.11.0",
    "pre-commit>=4.5.0",
    "pydantic-ai>=1.22.0",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.6",
    "streamlit>=1.51.0",
    "tavily-python>=0.7.13",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = [
  "F401"  # unused-import
]
